requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-identity==1.15.0
openai>=1.6.1,<2.0.0
langgraph>=0.0.20
langchain>=0.0.340
langchain-openai>=0.0.2
python-dotenv==1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0
requests>=2.31.0
pillow>=10.1.0
PyPDF2>=3.0.1
google-search-results>=2.4.2
pandas>=2.0.0
openpyxl>=3.1.0
pdf2image>=1.16.0
